Optimize API

+++ test_api.py
@@ -5,6 +5,7 @@ import argparse
 import tqdm
 import json
 import os
+import socket
 import sys
 from IPython.display import clear_output
 from pathlib import Path
@@ -15,14 +16,19 @@ if str(ROOT) not in sys.path:
     sys.path.append(str(ROOT))  # add ROOT to PATH
 ROOT = Path(os.path.abspath(ROOT / Path.cwd()))  # relative
 
+hostname = socket.gethostname()
+IP_ADDRESS = socket.gethostbyname(hostname)
+
 
 def parse_arg():
     parser = argparse.ArgumentParser()
-    parser.add_argument('--local_host', type=str, help='your local host connection', default='0.0.0.0')
+    parser.add_argument('--local_host', type=str, help='your local host connection', default=IP_ADDRESS)
     parser.add_argument('--port', type=str, help='your port connection', default='8000')
-    parser.add_argument('--json_results', type=str, help='your port connection', default=ROOT / 'results/results.json')
-    parser.add_argument('--folder', type=str, help='folder image test', default=ROOT / 'image')
-    parser.add_argument('--image', type=str, help='image path', default=os.path.join(ROOT, 'image/long.jpg'))
+    parser.add_argument('--json_results', type=str, help='your json path results',
+                        default=ROOT / 'results/results.json')
+    parser.add_argument('--source', type=str, help='folder image test', default=ROOT / 'image')
+    # parser.add_argument('--folder', type=str, help='folder image test', default=ROOT / 'image')
+    # parser.add_argument('--image', type=str, help='image path', default=os.path.join(ROOT, 'image/long.jpg'))
     return parser.parse_args()
 
 
@@ -30,13 +36,13 @@ def main():
     args = parse_arg()
     response = {}
     listResults = []
-    url = 'http://' + args.local_host + ':' + args.port + '/id-card-yolo/detect/'
+    url = 'http://' + args.local_host + ':' + args.port + '/id-card-yolo/detect_label_studio/'
     # with open(args.image, "rb") as image_file:
     #     data = base64.b64encode(image_file.read()).decode('utf-8')
     option = input('Do you want to show encoded image[Y/N]: ')
-    if args.folder:
-        for image in tqdm.tqdm(os.listdir(args.folder), total=len(args.folder)):
-            file = {'image': open(os.path.join(args.folder, image), 'rb')}
+    if os.path.isdir(args.source):
+        for image in tqdm.tqdm(os.listdir(args.source), total=len(args.source)):
+            file = {'image': open(os.path.join(args.source, image), 'rb')}
             data = {'option': option}
             response = requests.post(url=url, files=file, params=data)
             clear_output(wait=True)
@@ -44,8 +50,8 @@ def main():
             print(response.json(), '\n')
             with open(args.json_results, 'w') as fileSave:
                 json.dump(listResults, fileSave, indent=4)
-    else:
-        file = {'image': open(args.image, 'rb')}
+    elif os.path.isfile(args.source):
+        file = {'image': open(args.source, 'rb')}
         data = {'option': option}
         response = requests.post(url=url, files=file, params=data)
         with open(args.json_results, 'w') as fileSave:
